add validation logic to expenses router

+++ controllers/expense-controller.js
@@ -1,6 +1,7 @@
 const UserModel = require("../models/user-model");
 const ExpenseModel = require("./../models/expense-model");
 const errorCreator = require("../utils/error-creator");
+const { validationResult } = require("express-validator");
 
 exports.getAllExpenses = async (req, res, next) => {
   try {
@@ -26,6 +27,15 @@ exports.getAllExpenses = async (req, res, next) => {
 };
 
 exports.postAddNewExpense = async (req, res, next) => {
+  const validationErrors = validationResult(req);
+  if (!validationErrors.isEmpty()) {
+    const errorObject = validationErrors.array().map((obj) => {
+      return {
+        [obj.path]: obj.msg,
+      };
+    });
+    return next(errorCreator(errorObject, 400));
+  }
   const { date, amount, paymentMode, reason } = req.body;
   const userId = req.userId;
   const newExpense = new ExpenseModel({
@@ -53,13 +63,27 @@ exports.postAddNewExpense = async (req, res, next) => {
 exports.deleteRemoveExpense = async (req, res, next) => {
   const expenseId = req.params.expenseId;
   const userId = req.userId;
+  const validationErrors = validationResult(req);
+  if (!validationErrors.isEmpty()) {
+    const errorObject = validationErrors.array().map((obj) => {
+      return {
+        [obj.path]: obj.msg,
+      };
+    });
+    return next(errorCreator(errorObject, 400));
+  }
   try {
     const deletedExpense = await ExpenseModel.findOneAndDelete({
       _id: expenseId,
       user: userId,
     });
     if (!deletedExpense) {
-      return next(errorCreator(`Did not find an expense with the id = ${expenseId} belonging to the user ${userId}`, 404));
+      return next(
+        errorCreator(
+          `Did not find an expense with the id = ${expenseId} belonging to the user ${userId}`,
+          404
+        )
+      );
     }
     const updatedUser = await UserModel.findByIdAndUpdate(userId, {
       $pull: { expenses: deletedExpense._id },
@@ -76,6 +100,15 @@ exports.deleteRemoveExpense = async (req, res, next) => {
 };
 
 exports.putUpdateExpense = async (req, res, next) => {
+  const validationErrors = validationResult(req);
+  if (!validationErrors.isEmpty()) {
+    const errorObject = validationErrors.array().map((obj) => {
+      return {
+        [obj.path]: obj.msg,
+      };
+    });
+    return next(errorCreator(errorObject, 400));
+  }
   const userId = req.userId;
   const validProperties = ["paymentMode", "amount", "date", "reason"];
   const expenseId = req.params.expenseId;
@@ -92,7 +125,9 @@ exports.putUpdateExpense = async (req, res, next) => {
       { runValidators: true, returnDocument: "after" }
     );
     if (!updatedExpense) {
-      return next(errorCreator(`No expense found with the expenseId = ${expenseId}`, 404));
+      return next(
+        errorCreator(`No expense found with the expenseId = ${expenseId}`, 404)
+      );
     }
     res.status(200).json({
       status: "success",
